+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Yas Change History
 
-## 0.0.1
+## 1.0
 
-Initial release.
+Initial release. contains the following functionality:
+1) Adding an alias for a shell command.
+2) Remove an aliased command.
+3) Drop all aliased commands.
+4) List all aliased commands.
+5) Execute an aliased command.